refactor(cli): varredura de binarios do update sai para update_scan.rs
O #1030 levou update.rs a 716 linhas e o ratchet de qualidade acusou
(`files_over_700` 60 -> 61). A varredura do PATH e uma responsabilidade
propria — acha os outros garraia/garra, sonda a versao, formata o aviso
— e vai para o modulo `update_scan`, com os tres testes dela. update.rs
(442 linhas) fica com os pontos de entrada, que so chamam
`report_other_binaries`. Nada de comportamento muda.
